+++ codegen/src/attribute.rs
@@ -213,39 +213,60 @@ impl Step {
         if is_regex_or_expr {
             if let Some(elem_ty) = find_first_slice(&func.sig) {
                 let addon_parsing = Some(quote! {
-                    let __cucumber_matches = __cucumber_ctx
+                    let mut __cucumber_iter = __cucumber_ctx
                         .matches
                         .iter()
                         .skip(1)
-                        .enumerate()
-                        .map(|(i, (cap_name, s))| {
-                            // Special handling of `cucumber-expressions`
-                            // `string` parameter: removes leading and trailing
-                            // quotes.
-                            // See https://bit.ly/35XkDsp for more info.
-                            let is_string_group = cap_name
-                                .as_ref()
-                                .map(|n| n.starts_with("string_"))
-                                .unwrap_or_default();
+                        .enumerate();
 
-                            let s = if is_string_group && s.starts_with("\"") {
-                                s.strip_prefix("\"")
-                                    .and_then(|s| s.strip_suffix("\""))
-                                    .unwrap_or(s)
-                            } else if is_string_group && s.starts_with("'") {
-                                s.strip_prefix("'")
-                                    .and_then(|s| s.strip_suffix("'"))
-                                    .unwrap_or(s)
-                            } else {
-                                s
-                            };
+                    let mut __cucumber_matches = Vec::new();
+                    while let Some((i, (cap_name, s))) =
+                        __cucumber_iter.next()
+                    {
+                        // Special handling of `cucumber-expressions`
+                        // `parameter` with multiple capturing groups.
+                        let prefix = cap_name
+                            .as_ref()
+                            .filter(|n| n.starts_with("__"))
+                            .map(|n| {
+                                let num_len = n
+                                    .chars()
+                                    .skip(2)
+                                    .take_while(|&c| c != '_')
+                                    .map(char::len_utf8)
+                                    .sum::<usize>();
+                                let len = num_len + b"__".len();
+                                n.split_at(len).0
+                            });
 
+                        let to_take = __cucumber_iter
+                            .clone()
+                            .take_while(|(_, (n, _))| {
+                                prefix
+                                    .zip(n.as_ref())
+                                    .filter(|(prefix, n)| n.starts_with(prefix))
+                                    .is_some()
+                            })
+                            .count();
+
+                        let s = std::iter::once(s.as_str())
+                            .chain(
+                                __cucumber_iter
+                                    .by_ref()
+                                    .take(to_take)
+                                    .map(|(_, (_, s))| s.as_str()),
+                            )
+                            .fold(None, |acc, s| {
+                                acc.or_else(|| (!s.is_empty()).then(|| s))
+                            })
+                            .unwrap_or("");
+                        __cucumber_matches.push(
                             s.parse::<#elem_ty>().unwrap_or_else(|e| panic!(
                                 "Failed to parse element at {} '{}': {}",
                                 i, s, e,
                             ))
-                        })
-                        .collect::<Vec<_>>();
+                        );
+                    }
                 });
                 let func_args = func
                     .sig
@@ -341,29 +362,46 @@ impl Step {
 
             quote! {
                 let #ident = {
-                    // Special handling of `cucumber-expressions` `string`
-                    // parameter: removes leading and trailing quotes.
-                    // See https://bit.ly/35XkDsp for more info.
                     let (cap_name, s) = __cucumber_iter
                         .next()
                         .expect(#not_found_err);
-
-                    let is_string_group = cap_name
+                    // Special handling of `cucumber-expressions` `parameter`
+                    // with multiple capturing groups.
+                    let prefix = cap_name
                         .as_ref()
-                        .map(|n| n.starts_with("string_"))
-                        .unwrap_or_default();
+                        .filter(|n| n.starts_with("__"))
+                        .map(|n| {
+                            let num_len = n
+                                .chars()
+                                .skip(2)
+                                .take_while(|&c| c != '_')
+                                .map(char::len_utf8)
+                                .sum::<usize>();
+                            let len = num_len + b"__".len();
+                            n.split_at(len).0
+                        });
 
-                    if is_string_group && s.starts_with("\"") {
-                        s.strip_prefix("\"")
-                            .and_then(|s| s.strip_suffix("\""))
-                            .unwrap_or(s)
-                    } else if is_string_group && s.starts_with("'") {
-                        s.strip_prefix("'")
-                            .and_then(|s| s.strip_suffix("'"))
-                            .unwrap_or(s)
-                    } else {
-                        s
-                    }
+                    let to_take = __cucumber_iter
+                        .clone()
+                        .take_while(|(n, _)| {
+                            prefix.zip(n.as_ref())
+                                .filter(|(prefix, n)| n.starts_with(prefix))
+                                .is_some()
+                        })
+                        .count();
+
+                    std::iter::once(s.as_str())
+                        .chain(
+                            __cucumber_iter
+                                .by_ref()
+                                .take(to_take)
+                                .map(|(_, s)| s.as_str())
+                        )
+                        .fold(
+                            None,
+                            |acc, s| acc.or_else(|| (!s.is_empty()).then(|| s)),
+                        )
+                        .unwrap_or("")
                 };
                 let #ident = #ident.parse::<#ty>().expect(#parsing_err);
             }