+++ codegen/src/lib.rs
@@ -309,7 +309,8 @@ steps!(given, when, then);
 ///
 /// - `#[param(regex = "regex")]`
 ///
-///   [`Regex`] to match this parameter. Shouldn't contain any capturing groups.
+///   [`Regex`] to match this parameter. Usually shouldn't contain any capturing
+///   groups, but in case it does, only first non-empty group will be matched.
 ///
 /// - `#[param(name = "name")]` (optional)
 ///